+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/TileEntitySqueezer.java
@@ -156,6 +156,7 @@ public class TileEntitySqueezer extends TileEntityMultiblockMetal<TileEntitySque
 				{
 					ItemStack full = Utils.fillFluidContainer(tanks[0], empty, getInventory().get(10), null);
 					if(!full.isEmpty())
+					{
 						if(getInventory().get(9).getCount()==1&&!Utils.isFluidContainerFull(full))
 							getInventory().set(9, full.copy());
 						else
@@ -168,6 +169,8 @@ public class TileEntitySqueezer extends TileEntityMultiblockMetal<TileEntitySque
 							if(inventory.get(9).getCount() <= 0)
 								inventory.set(9, ItemStack.EMPTY);
 						}
+						update = true;
+					}
 				}
 			}
 			if(!inventory.get(8).isEmpty() && world.getTotalWorldTime()%8==0)